Phase B Batch 2: 4-plane first-hit restructure
10-impl-b.md Batch 2 section. Restructures first-hit into NAADF's
base/ pipeline shape: atmosphere -> 4-plane first-hit -> final-blit
chain dispatches every frame. The A-2 taa_samples write is removed
from the albedo path (designed seam, item #1 accepted by user);
blit temporarily points at final_color, Batch 6 rewires it into
base/ TAA. Item #2 resolved: A-2 has no latent bug -- albedo source
genuinely says screenPosDistanceSqr > 1.0, base/ genuinely says
16.0; pipelines differ, no erratum needed. Fixed a latent Batch-1
bug (atmosphere_oct_index missing explicit u32() casts). 41 tests
pass, smoke-run clean, visual gate confirmed (emissive white,
others black, sky rendering) -> atmosphere-visible.

+++ src/assets/shaders/atmosphere.wgsl
@@ -274,9 +274,12 @@ fn atmosphere_oct_index(
     ray_dir.z *= xz_scale;
 
     let oct = oct_encode(ray_dir);
+    // HLSL `uint2(oct.x * (texSizeX-1), ...)` implicitly truncates float→uint
+    // in the `uint2(...)` constructor; WGSL requires explicit `u32()` casts
+    // (naga rejects building a `vec2<u32>` from `f32` components).
     let comp_pos = vec2<u32>(
-        oct.x * f32(atmosphere_tex_size_x - 1u),
-        oct.y * f32(atmosphere_tex_size_y - 1u),
+        u32(oct.x * f32(atmosphere_tex_size_x - 1u)),
+        u32(oct.y * f32(atmosphere_tex_size_y - 1u)),
     );
     return comp_pos.x + comp_pos.y * atmosphere_tex_size_x;
 }
